feat: equal 버튼 클릭 시 계산되는 함수 calculator로 이동
