+++ src/components/OperatorButton.js
@@ -1,12 +1,8 @@
 import React from 'react';
-import {
-  OPERATOR,
-  ERROR_MSG,
-  CALCULATOR_INITIAL_STATE,
-} from '../constants/constant';
+import { OPERATOR, CALCULATOR_INITIAL_STATE } from '../constants/constant';
 
-const OperatorButton = ({ selfOperand, state, setState }) => {
-  const { prevNumbers, operator, nextNumbers } = state;
+const OperatorButton = ({ selfOperand, state, setState, calculate }) => {
+  const { prevNumbers } = state;
 
   const onClickOperator = () => {
     if (prevNumbers.length === 0) {
@@ -22,31 +18,7 @@ const OperatorButton = ({ selfOperand, state, setState }) => {
       return;
     }
 
-    setState({
-      ...CALCULATOR_INITIAL_STATE,
-      sum: calculateSum(),
-    });
-  };
-
-  const calculateSum = () => {
-    const prevNumber = Number(prevNumbers.join(''));
-    const nextNumber = Number(nextNumbers.join(''));
-
-    switch (operator) {
-      case OPERATOR.PLUS:
-        return prevNumber + nextNumber;
-      case OPERATOR.SUBSTRACT:
-        return prevNumber - nextNumber;
-      case OPERATOR.MULTI:
-        return prevNumber * nextNumber;
-      case OPERATOR.DIVIDE:
-        if (!isFinite(prevNumber / nextNumber)) {
-          return ERROR_MSG.INFINITY;
-        }
-        return Math.floor(prevNumber / nextNumber);
-      default:
-        return prevNumber;
-    }
+    calculate();
   };
 
   return (
